+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [NEXT] - YYYY-MM-DD
 
-- Allows Elytra flight to start and stop client-side. (1.19.2)
+- Allows Elytra Flight to start and stop client-side. (1.19.2)
+- Allows using a firework client-side when in Elytra Flight. (1.19.2)
 
 ## [0.0.6] - 2023-05-02